Fix an issue when two triangle share 2 vertices that are on the plane.
  2
 /  \
1 - -3
 \  /
   4

In this situation, the previous implementation could start from triangle
(1, 2, 3), visit vertex 1, use the triangle (1, 3, 4) as the next triangle
and then visit vertex 3 and pick the initial triangle again, at which
point it would stop thinking it has been around the mesh.

This issue happen only when two triangles share two vertices that lie
exactly on the plane (so both triangles have the same set of intersections
with the plane).

The solution is that once we have visited a triangle and chosed the next
one from its neighbours, we remove all neighbouring triangles from the
‘to-visit’ list. This ensure we never go back to a previously visited
triangle.

Closes #7